changed int input to float input

+++ average.py
@@ -4,11 +4,11 @@
 # Python Project 2a: Average of 5 input numbers
 
 
-num_1 = int(input("Please Enter 5 numbers, Separated by Enter: "))
-num_2 = int(input())
-num_3 = int(input())
-num_4 = int(input())
-num_5 = int(input())
+num_1 = float(input("Please Enter 5 numbers, Separated by Enter: "))
+num_2 = float(input())
+num_3 = float(input())
+num_4 = float(input())
+num_5 = float(input())
 sum = num_1 + num_2 + num_3 + num_4 + num_5
 print("the average of those numbers is")
-print(int(sum / 5))
+print(float(sum / 5))
